✨ feat: icon 동적 import한 ICON_MAP 상수 폴더로 분리
Ref: [feat] 아이콘 컴포넌트 구현 #8
