+++ src/constants/iconMap.ts
@@ -1,0 +1,39 @@
+const ICON_MAP = {
+  AlarmActive: () => import('@/assets/icons/alarm_active.svg'),
+  AlarmInActive: () => import('@/assets/icons/alarm_inactive.svg'),
+  ArrowLeft: () => import('@/assets/icons/arrow_left.svg'),
+  ArrowRight: () => import('@/assets/icons/arrow_right.svg'),
+  Bus: () => import('@/assets/icons/bus.svg'),
+  Calender: () => import('@/assets/icons/calender.svg'),
+  ChevronLeft: () => import('@/assets/icons/chevron-left.svg'),
+  ChevronRight: () => import('@/assets/icons/chevron-right.svg'),
+  Delete: () => import('@/assets/icons/delete.svg'),
+  Edit: () => import('@/assets/icons/edit.svg'),
+  EyeOff: () => import('@/assets/icons/eye_off.svg'),
+  EyeOn: () => import('@/assets/icons/eye_on.svg'),
+  Facebook: () => import('@/assets/icons/facebook.svg'),
+  Food: () => import('@/assets/icons/food.svg'),
+  Instagram: () => import('@/assets/icons/instagram.svg'),
+  List: () => import('@/assets/icons/list.svg'),
+  Map: () => import('@/assets/icons/map.svg'),
+  Minus: () => import('@/assets/icons/minus.svg'),
+  More: () => import('@/assets/icons/more.svg'),
+  Music: () => import('@/assets/icons/art.svg'),
+  Plus: () => import('@/assets/icons/plus.svg'),
+  Search: () => import('@/assets/icons/search.svg'),
+  Setting: () => import('@/assets/icons/setting.svg'),
+  Sport: () => import('@/assets/icons/sport.svg'),
+  Star: () => import('@/assets/icons/star.svg'),
+  Tour: () => import('@/assets/icons/tour.svg'),
+  TriangleDown: () => import('@/assets/icons/triangle_down.svg'),
+  TriangleLeft: () => import('@/assets/icons/triangle_left.svg'),
+  TriangleRight: () => import('@/assets/icons/triangle_right.svg'),
+  TriangleUp: () => import('@/assets/icons/triangle_up.svg'),
+  User: () => import('@/assets/icons/user.svg'),
+  Warning: () => import('@/assets/icons/warning.svg'),
+  Wellbeing: () => import('@/assets/icons/wellbeing.svg'),
+  X: () => import('@/assets/icons/X.svg'),
+  Youtube: () => import('@/assets/icons/youtube.svg'),
+};
+
+export default ICON_MAP;
